Refactor Elements
